Added basic reflect system
Pressing "r" will return a unit to the start of their turn. Units can only move right now, so this will need to be updated. I'll probably have each element of the actionStack be a dictionary that notes what the tile was.
I'm going to implement unit selection next since I feel like it's important for the actual puzzle components and unit actions.